Module anoma_tx_prelude::proof_of_stake::vp

Proof-of-Stake native validity predicate.

Modules

anoma_proof_of_stake

Proof of Stake system.

Structs

PosParams

Proof-of-Stake system parameters

PosVP

Proof-of-Stake validity predicate

Slash

A slash applied to validator, to punish byzantine behavior by removing their staked tokens at and before the epoch of the slash.

Enums

Error

Type Definitions

Result

PoS functions result

Slashes

Slashes applied to validator, to punish byzantine behavior by removing their staked tokens at and before the epoch of the slash.

TotalVotingPowers

Epoched total voting power.

ValidatorStates

Epoched validator’s state.

ValidatorVotingPowers

Epoched validator’s voting power.