Module anoma_tx_prelude::storage

governance storage

Functions

get_author_key

Get key of proposal author

get_committing_proposals_key

Get the committing proposal key

get_content_key

Get key of proposal content

get_counter_key

Get key of proposal ids counter

get_funds_key

Get key of proposal funds

get_grace_epoch_key

Get proposal grace epoch key

get_max_proposal_code_size_key

Get maximum proposal code size key

get_max_proposal_content_key

Get maximum proposal content key

get_min_proposal_fund_key

Get key for the minimum proposal fund

get_min_proposal_grace_epoch_key

Get min grace epoch proposal key

get_min_proposal_period_key

Get minimum proposal period key

get_proposal_code_key

Get proposal code key

get_voting_end_epoch_key

Get key of proposal voting end epoch

get_voting_start_epoch_key

Get key of proposal voting start epoch

is_author_key

Check if key is author key

is_balance_key

Check if key is balance key

is_commit_proposal_key

Check if key is a commit proposal key

is_content_key

Check if key is content key

is_counter_key

Check if key is counter key

is_end_epoch_key

Check if key is epoch key

is_governance_key

Check if key is inside governance address space

is_grace_epoch_key

Check if key is grace epoch key

is_max_content_size_key

Check if key is a proposal max content parameter key

is_max_proposal_code_size_key

Check if key is a max proposal size key

is_min_grace_epoch_key

Check if key is a min grace epoch key

is_min_proposal_fund_key

Check if key is a proposal fund parameter key

is_min_proposal_period_key

Check if key is a min proposal period param key

is_parameter_key

Check if key is parameter key

is_proposal_code_key

Check if key is proposal key

is_start_epoch_key

Check if key is start epoch key

is_start_or_end_epoch_key

Check if key is start epoch or end epoch key

is_vote_key

Check if a key is a vote key

proposal_prefix

Get governance prefix key