Module anoma_tx_prelude::time

Types for dealing with time and durations.

Structs

DateTime

ISO 8601 combined date and time with time zone.

DateTimeUtc

A duration in seconds precision.

Duration

ISO 8601 time duration with nanosecond precision. This also allows for the negative duration; see individual methods for details.

DurationNanos

A duration in nanos precision.

DurationSecs

A duration in seconds precision.

Rfc3339String

An RFC 3339 timestamp (e.g., “1970-01-01T00:00:00Z”).

Utc

The UTC time zone. This is the most efficient time zone when you don’t need the local time. It is also used as an offset (which is also a dummy type).

Traits

TimeZone

The time zone.

Functions

duration_passed

Check if the given duration has passed since the given `start.