Trait anoma_tx_prelude::PosWrite

pub trait PosWrite: PosReadOnly {

    fn write_pos_params(&mut self, params: &PosParams);

    fn write_validator_address_raw_hash(&mut self, address: &Self::Address);

    fn write_validator_staking_reward_address(
        &mut self, 
        key: &Self::Address, 
        value: Self::Address
    );

    fn write_validator_consensus_key(
        &mut self, 
        key: &Self::Address, 
        value: Epoched<Self::PublicKey, OffsetPipelineLen>
    );

    fn write_validator_state(
        &mut self, 
        key: &Self::Address, 
        value: Epoched<ValidatorState, OffsetPipelineLen>
    );

    fn write_validator_total_deltas(
        &mut self, 
        key: &Self::Address, 
        value: EpochedDelta<Self::TokenChange, OffsetUnboundingLen>
    );

    fn write_validator_voting_power(
        &mut self, 
        key: &Self::Address, 
        value: EpochedDelta<VotingPowerDelta, OffsetUnboundingLen>
    );

    fn write_bond(
        &mut self, 
        key: &BondId<Self::Address>, 
        value: EpochedDelta<Bond<Self::TokenAmount>, OffsetPipelineLen>
    );

    fn write_unbond(
        &mut self, 
        key: &BondId<Self::Address>, 
        value: EpochedDelta<Unbond<Self::TokenAmount>, OffsetUnboundingLen>
    );

    fn write_validator_set(
        &mut self, 
        value: Epoched<ValidatorSet<Self::Address>, OffsetUnboundingLen>
    );

    fn write_total_voting_power(
        &mut self, 
        value: EpochedDelta<VotingPowerDelta, OffsetUnboundingLen>
    );

    fn delete_bond(&mut self, key: &BondId<Self::Address>);

    fn delete_unbond(&mut self, key: &BondId<Self::Address>);

    fn transfer(
        &mut self, 
        token: &Self::Address, 
        amount: Self::TokenAmount, 
        src: &Self::Address, 
        dest: &Self::Address
    );

    fn become_validator(
        &mut self, 
        address: &Self::Address, 
        staking_reward_address: &Self::Address, 
        consensus_key: &Self::PublicKey, 
        current_epoch: impl Into<Epoch>
    ) -> Result<(), BecomeValidatorError<Self::Address>> { ... }

    fn is_validator(&self, address: &Self::Address) -> bool { ... }

    fn bond_tokens(
        &mut self, 
        source: Option<&Self::Address>, 
        validator: &Self::Address, 
        amount: Self::TokenAmount, 
        current_epoch: impl Into<Epoch>
    ) -> Result<(), BondError<Self::Address>> { ... }

    fn unbond_tokens(
        &mut self, 
        source: Option<&Self::Address>, 
        validator: &Self::Address, 
        amount: Self::TokenAmount, 
        current_epoch: impl Into<Epoch>
    ) -> Result<(), UnbondError<Self::Address, Self::TokenAmount>> { ... }

    fn withdraw_tokens(
        &mut self, 
        source: Option<&Self::Address>, 
        validator: &Self::Address, 
        current_epoch: impl Into<Epoch>
    ) -> Result<Self::TokenAmount, WithdrawError<Self::Address>> { ... }

}

PoS system trait to be implemented in integration that can read and write PoS data.

Required methods

fn write_pos_params(&mut self, params: &PosParams)

Write PoS parameters.

fn write_validator_address_raw_hash(&mut self, address: &Self::Address)

Write PoS validator’s raw hash its address.

fn write_validator_staking_reward_address(
&mut self,
key: &Self::Address,
value: Self::Address
)

Write PoS validator’s staking reward address, into which staking rewards will be credited.

fn write_validator_consensus_key(
&mut self,
key: &Self::Address,
value: Epoched<Self::PublicKey, OffsetPipelineLen>
)

Write PoS validator’s consensus key (used for signing block votes).

fn write_validator_state(
&mut self,
key: &Self::Address,
value: Epoched<ValidatorState, OffsetPipelineLen>
)

Write PoS validator’s state.

fn write_validator_total_deltas(
&mut self,
key: &Self::Address,
value: EpochedDelta<Self::TokenChange, OffsetUnboundingLen>
)

Write PoS validator’s total deltas of their bonds (validator self-bonds and delegations).

fn write_validator_voting_power(
&mut self,
key: &Self::Address,
value: EpochedDelta<VotingPowerDelta, OffsetUnboundingLen>
)

Write PoS validator’s voting power.

fn write_bond(
&mut self,
key: &BondId<Self::Address>,
value: EpochedDelta<Bond<Self::TokenAmount>, OffsetPipelineLen>
)

Write PoS bond (validator self-bond or a delegation).

fn write_unbond(
&mut self,
key: &BondId<Self::Address>,
value: EpochedDelta<Unbond<Self::TokenAmount>, OffsetUnboundingLen>
)

Write PoS unbond (unbonded tokens from validator self-bond or a delegation).

fn write_validator_set(
&mut self,
value: Epoched<ValidatorSet<Self::Address>, OffsetUnboundingLen>
)

Write PoS validator set (active and inactive).

fn write_total_voting_power(
&mut self,
value: EpochedDelta<VotingPowerDelta, OffsetUnboundingLen>
)

Write PoS total voting power of all validators (active and inactive).

fn delete_bond(&mut self, key: &BondId<Self::Address>)

Delete an emptied PoS bond (validator self-bond or a delegation).

fn delete_unbond(&mut self, key: &BondId<Self::Address>)

Delete an emptied PoS unbond (unbonded tokens from validator self-bond or a delegation).

fn transfer(
&mut self,
token: &Self::Address,
amount: Self::TokenAmount,
src: &Self::Address,
dest: &Self::Address
)

Transfer tokens from the src to the dest.

Provided methods

fn become_validator(
&mut self,
address: &Self::Address,
staking_reward_address: &Self::Address,
consensus_key: &Self::PublicKey,
current_epoch: impl Into<Epoch>
) -> Result<(), BecomeValidatorError<Self::Address>>

Attempt to update the given account to become a validator.

fn is_validator(&self, address: &Self::Address) -> bool

Check if the given address is a validator by checking that it has some state.

fn bond_tokens(
&mut self,
source: Option<&Self::Address>,
validator: &Self::Address,
amount: Self::TokenAmount,
current_epoch: impl Into<Epoch>
) -> Result<(), BondError<Self::Address>>

Self-bond tokens to a validator when source is None or equal to the validator address, or delegate tokens from the source to the validator.

fn unbond_tokens(
&mut self,
source: Option<&Self::Address>,
validator: &Self::Address,
amount: Self::TokenAmount,
current_epoch: impl Into<Epoch>
) -> Result<(), UnbondError<Self::Address, Self::TokenAmount>>

Unbond self-bonded tokens from a validator when source is None or equal to the validator address, or unbond delegated tokens from the source to the validator.

fn withdraw_tokens(
&mut self,
source: Option<&Self::Address>,
validator: &Self::Address,
current_epoch: impl Into<Epoch>
) -> Result<Self::TokenAmount, WithdrawError<Self::Address>>

Withdraw unbonded tokens from a self-bond to a validator when source is None or equal to the validator address, or withdraw unbonded tokens delegated to the validator to the source.